Potential fix for code scanning alert no. 53: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to prevent overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done in the larger type, and the result will be correctly calculated without overflow.

Specifically, we will cast the first operand 2 to unsigned long before performing the multiplication. This change will be made on line 72 of the file libwebp-1.0.3/src/utils/rescaler_utils.c.

+++ libwebp-1.0.3/src/utils/rescaler_utils.c
@@ -69,7 +69,7 @@ void WebPRescalerInit(WebPRescaler* const wrk, int src_width, int src_height,
   }
   wrk->irow = work;
   wrk->frow = work + num_channels * dst_width;
-  memset(work, 0, 2 * dst_width * num_channels * sizeof(*work));
+  memset(work, 0, (unsigned long)2 * dst_width * num_channels * sizeof(*work));
 
   WebPRescalerDspInit();
 }
